fix the test and linters errors
