+++ spec/book_spec.rb
@@ -1,30 +1,33 @@
-require_relative '../book_label/book
+require_relative '../classes/book'
 
 describe Book do
   before :all do
-    @book = Book.new('1998-03-15", 'fantasy', 'good')
+    @book = Book.new('1998-03-15', 'Jenny', 'good')
   end
 
-context 'when creating a new book' do
-  it 'should be an instance of book class' do
-    expect(@book).to be_an_instance_of(Book)
-end
+  context 'when creating a new book' do
+    it 'should be an instance of book class' do
+      expect(@book).to be_an_instance_of(Book)
+    end
 
+    it 'should have a cover state' do
+      expect(@book.cover_state).to eq('good')
+    end
 
-it 'should have a cover state' do
-  expect(@book.cover_state).to eq('good')
-end
+    it 'should have a publisher' do
+      expect(@book.publisher).to eq('Jenny')
+    end
 
-it 'should have a publisher' do
-  expect(@book.publisher).to eq('fantasy')
-end
+    it 'should have a publish date' do
+      expect(@book.publish_date.year).to eq(1998)
+    end
 
-it 'should have a published date' do
-  expect(@book.published_date).to eq(1998-03-15')
-end
+    it 'has can_be_archived? method' do
+      expect(@book).to respond_to(:can_be_archived?)
+    end
 
-it 'has can_be_archived? method' do
-   expect(@book).to respond_to(:can_be_archived?)
+    it 'has move_to_archive method' do
+      expect(@book.move_to_archive).to eq(true)
+    end
   end
- end
 end
